feat(settings): unify notifications into one event x channel matrix
Per-environment notification settings were two separate axes (a channel
list + an event-type list, each inherit/on/off), while the global
settings were an event x channel matrix. That was inconsistent and
actually less expressive than global — you couldn't override a single
event+channel for an environment.

Unify both scopes on one matrix:
- Extract a reusable <NotificationMatrix> component. Global cells are
  on/off switches; per-environment cells are inherit/on/off, where
  "inherit" falls back to the global cell for that exact event+channel
  (environment + event + channel is the most specific resolver scope).
- Settings (global + per-environment expandable) and the environment
  notification modal all render the same matrix.
- Composable: replace the coarse env channel/event helpers
  (envChannelState/setEnvChannel/envEventState/setEnvEvent) with
  envEventChannelState/setEnvEventChannel.

No backend change; the resolver already supported this granularity.
